% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rofanova}
\alias{rofanova}
\title{Robust Functional Analysis of Variance}
\usage{
rofanova(
  X,
  label_1,
  label_2 = NULL,
  B = 100,
  cores = 1,
  family = "bisquare",
  eff = 0.95,
  mu0_g = NULL,
  scale = NULL,
  maxit = 50,
  tol = 1e-04
)
}
\arguments{
\item{X}{Either an object of class  \code{fdata} for monodimensional functional data  or an object of class \code{fdata2d} for bi-dimensional functional data.}

\item{label_1}{A vector of containing group label corresponding to the first main effect.}

\item{label_2}{A vector of containing group label corresponding to the first main effect. If it is NULL, the one-way RoFANOVA is performed.
Otherwise, the two-way RoFANOVA with interaction is employed. Default is NULL.}

\item{B}{The number of permutations used to approximate the p-value in the permutation test. Default is 1000.}

\item{cores}{If \code{cores}>1, then parallel computing is used, with \code{cores} cores. Default is 1.
@param family The family of loss function for the calcualtion of the equivariant functional M-estimator. The values allowed are
"bisquare" for the bisquare or Tukey's biweight family of loss functions;  "huber" for the the Huber's family of loss functions;
  "optimal" for the  optimal family of loss functions; "hampel" for the the Hampel's family of loss functions; "median" for the median loss function.
  A non-robust functional estimator of the mean based on the standard least squares loss function is used with the value "mean". Default is "bisquare".}

\item{eff}{Asymptotic efficiency of the equivariant functional M-estimator. When \code{family} is either "mean" or "median", \code{eff} is ignored.}

\item{mu0_g}{Initial estimate  used in re-weighted least-squares algorithm to compute the equivariant functional M-estimator.
If NULL the standard non-robust functional mean is used. Default is NULL.}

\item{scale}{Estimate of the standard error of \code{X}. If NUll, the functional normalized median absolute deviation estimator is used. Default is NULL.}

\item{maxit}{The maximum number of iterations allowed in the re-weighted least-squares algorithm to compute the equivariant functional M-estimator.}

\item{tol}{The tolerance for the stopping condition of the re-weighted least-squares algorithm to compute the equivariant functional M-estimator.
The algorithm stops when the relative variation of the weighted norm sum between two consecutive iterations is less than \code{tol}.}
}
\value{
A list containing the following arguments:
\code{mod} that is a list composed by
\itemize{
\item \code{data}: A list containing the vectorized form of \code{X}, \code{timeindex}, and \code{curve}. For functional data observed over a regular grid \code{timeindex} and \code{curve} are trivially obtained.

\item \code{parameters}: A list containing all the estimated parameters.

\item \code{vars}: A list containing results from the Expectation step of the ECM algorithm.

\item \code{FullS}: The matrix of B-spline computed over \code{grid}.

\item \code{grid}: The vector of time points where the curves are sampled.

\item \code{W}: The basis roughness penalty matrix containing the inner products of pairs of basis function second derivatives.

\item \code{AW_vec}: Vectorized version of the diagonal matrix used in the approximation of FAPFP.

\item \code{P_tot}: Sparse Matrix used to compute all the pairwise comparisons in the FAPFP.

\item \code{lambda_s}: Tuning parameter of the smoothness penalty.

\item \code{lambda_l}: Tuning parameter of the FAPFP.
}

A list, named \code{clus}, containing the following arguments:
\itemize{
\item \code{classes}: The vector of cluster membership.

\item \code{po_pr}: Posterior probabilities of cluster membership.
}

\code{mean_fd} The estimated cluster mean functions.

\code{class} A label for the output type.
}
\description{
Robust Functional Analysis of Variance (RoFANOVA) allows identifying the presence of significant differences, in terms of
functional mean, among groups of a functional data by being robust against the presence of outliers  (Centofanti et al., 2021).
}
\examples{
library(sasfunclust)
train<-simulate_data("Scenario I",n_i=20,var_e = 1,var_b = 0.5^2)
mod<-sasfclust(X=train$X,grid=train$grid,lambda_s = 10^-6,lambda_l =10,G = 2,maxit = 5,q=10)
plot(mod)
}
\references{
Centofanti, F., Lepore, A., & Palumbo, B. (2021).
Sparse and Smooth Functional Data Clustering.
\emph{arXiv preprint arXiv:2103.15224}.

Ramsay, J., Ramsay, J., & Silverman, B. W. (2005). Functional Data Analysis. Springer Science & Business Media.
}
\seealso{
\code{\link{sasfclust_cv}}
}
